refactor: split up tree and browsing logic
